requirements.txt
Django==4.2.7
gunicorn>=21.0.0
PyMySQL>=1.1.0
Pillow>=10.0.0
python-decouple==3.8
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
whitenoise>=6.6.0
